Reduced greatly increased efficiency of code (and it actually works now)
Previous code was actually broken, because the entire list pointed to
pt, instead of becoming a copy of it's value. Now, new arrays are
manually assigned per point (manual assignment came out to be the
fastest in tests). Also reduced the amount of overlap, and increased the
initial capacity of the arraylist (this actually helped quite a bit).